requirements.txt
python-telegram-bot==20.3
pandas
numpy
pandas_ta
requests
python-dotenv
SQLAlchemy
apscheduler
matplotlib
scipy
